Fix docker merge conflicts

Kindly done by @wiggin77 
Thank you :)

+++ server/app/category.go
@@ -1,9 +1,15 @@
 package app
 
 import (
+	"errors"
+	"strings"
+
 	"github.com/mattermost/focalboard/server/model"
 	"github.com/mattermost/focalboard/server/utils"
 )
+
+var ErrCannotDeleteSystemCategory = errors.New("cannot delete a system category")
+var ErrCannotUpdateSystemCategory = errors.New("cannot update a system category")
 
 func (a *App) CreateCategory(category *model.Category) (*model.Category, error) {
 	category.Hydrate()
@@ -28,6 +34,14 @@ func (a *App) CreateCategory(category *model.Category) (*model.Category, error) 
 }
 
 func (a *App) UpdateCategory(category *model.Category) (*model.Category, error) {
+	// set to default category, UI doesn't create with Type
+	if strings.TrimSpace(category.Type) == "" {
+		category.Type = model.CategoryTypeCustom
+	}
+	if err := category.IsValid(); err != nil {
+		return nil, err
+	}
+
 	// verify if category belongs to the user
 	existingCategory, err := a.store.GetCategory(category.ID)
 	if err != nil {
@@ -40,6 +54,19 @@ func (a *App) UpdateCategory(category *model.Category) (*model.Category, error) 
 
 	if existingCategory.UserID != category.UserID {
 		return nil, model.ErrCategoryPermissionDenied
+	}
+
+	if existingCategory.TeamID != category.TeamID {
+		return nil, model.ErrCategoryPermissionDenied
+	}
+
+	// in case type was defaulted above, set to existingCategory.Type
+	category.Type = existingCategory.Type
+	if existingCategory.Type == model.CategoryTypeSystem {
+		// You cannot rename or delete a system category,
+		// So restoring its name and undeleting it if set so.
+		category.Name = existingCategory.Name
+		category.DeleteAt = 0
 	}
 
 	category.UpdateAt = utils.GetMillis()
@@ -82,6 +109,10 @@ func (a *App) DeleteCategory(categoryID, userID, teamID string) (*model.Category
 	// verify if category belongs to the team
 	if existingCategory.TeamID != teamID {
 		return nil, model.NewErrInvalidCategory("category doesn't belong to the team")
+	}
+
+	if existingCategory.Type == model.CategoryTypeSystem {
+		return nil, ErrCannotDeleteSystemCategory
 	}
 
 	if err = a.store.DeleteCategory(categoryID, userID, teamID); err != nil {
